Radial Regression
For Teapot and Rotated Digits. Change of basis of the centered embedding to obtain angles phi as predictors for the ground truth angles. The smallest angles of both phi and ground truth angles are aligned to 0 to avoid issues with rotations. The returned value is the R^2 of the fitted regression. If more than one class is present (ex: RotatedDigits), the embedding is partioned according to class labels and the average R^2 over all partitions is returned.