Add classification metric to single cell datasets
The addition of a "partition" attribute allows to turn off the procedure whereby linear regressions are computed independently for each class in the data for some latent factor (think of this as separately probing the structure of different clusters). This allows to use the time labels of single cell datasets for both linear regression (since they won't be used to divide the data) and classification (since their domain is discrete).